pyproject.toml
[project]
name = "report-system"
version = "0.1.0"
description = ""
authors = [
    {name = "IvanArsenev",email = "arsenev.ivan00@bk.ru"}
]
readme = "README.md"
requires-python = "3.11"
dependencies = [
    "fastapi (>=0.116.0,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "aiogram (>=3.21.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
